Restructure navigations and test folders
