+++ src/__tests__/navigations/auth-navigator.test.tsx
@@ -1,5 +1,5 @@
 import { NavigationContainer } from '@react-navigation/native';
-import AuthNavigator from '../auth-navigator';
+import AuthNavigator from '@/navigations/auth-navigator';
 import { SignInFormTestId } from '@/features/auth/components/signin-form';
 import { SignInFormContainerTestId } from '@/features/auth/components/signin-form-container';
 import { SignUpFormTestId } from '@/features/auth/components/signup-form';
